Move disabled test to base class.
